Change to renderless component
