+++ v3/src/js/views/timetable/TimetableContainer.test.jsx
@@ -107,9 +107,9 @@ test('should display spinner when loading modules', () => {
 test('should display imported timetable', () => {
   const timetable = { CS2105: { Lecture: '1' } };
   const wrapper = createWithImport(timetable, { CS2105: CS3216 }).wrapper;
-
-  expect(wrapper.type()).toEqual(TimetableContent);
-  expect(wrapper.props()).toMatchObject({
+  const timetableContentWrapper = wrapper.children(TimetableContent);
+  expect(timetableContentWrapper.type()).toEqual(TimetableContent);
+  expect(timetableContentWrapper.props()).toMatchObject({
     timetable,
     readOnly: true,
   });
@@ -120,9 +120,9 @@ test('should display saved timetable when there is no imported timetable', () =>
   const timetable = { CS2105: { Lecture: '1' } };
 
   const wrapper = create(timetablePage(semester), { semester: 'sem-1' }, timetable).wrapper;
-
-  expect(wrapper.type()).toEqual(TimetableContent);
-  expect(wrapper.props()).toMatchObject({
+  const timetableContentWrapper = wrapper.children(TimetableContent);
+  expect(timetableContentWrapper.type()).toEqual(TimetableContent);
+  expect(timetableContentWrapper.props()).toMatchObject({
     timetable,
     readOnly: false,
   });
